frontend_chamazetu/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==41.0.7
deprecation==2.1.0
Django==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-sslserver==0.22
django-timezone-field==6.1.0
ecdsa==0.18.0
exceptiongroup==1.2.0
gotrue==2.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
kombu==5.3.5
numpy==1.26.4
packaging==23.2
pandas==2.2.1
postgrest==0.15.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.1
realtime==1.0.2
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
websockets==11.0.3
whitenoise==6.6.0
